More robust file locking

+++ smartdispatch/filelock.py
@@ -3,6 +3,7 @@ import time
 import fcntl
 import psutil
 import logging
+import errno
 
 from contextlib import contextmanager
 
@@ -36,12 +37,28 @@ def open_with_flock(*args, **kwargs):
     try:
         fcntl.lockf(f, fcntl.LOCK_EX | fcntl.LOCK_NB)
     except IOError:
-        logging.info("Can't immediately write-lock the file ({0}), waiting ...".format(f.name))
-        fcntl.lockf(f, fcntl.LOCK_EX)
+        no_attempt = 0
+        while no_attempt < MAX_ATTEMPTS:
+            try:
+                logging.info("Can't immediately write-lock the file ({0}), waiting.".format(f.name))
+                start_time = time.time()
+                fcntl.lockf(f, fcntl.LOCK_EX)
+                break
+            except IOError as e:
+                if e.errno == errno.EDEADLK:
+                    logging.warn("The OS complained because the process have been waiting on the lockf for {0} sec with the error ({1}: {2}). Retrying. ".format(time.time() - start_time), e.errno, e.strerror)
+                    no_attempt += 1
+                else:
+                    raise e
 
-    yield f
-    fcntl.lockf(f, fcntl.LOCK_UN)
-    f.close()
+        if no_attempt == MAX_ATTEMPTS:
+            raise IOError("Failed to lock {0} {1} times.".format(f.name, MAX_ATTEMPTS))
+
+    try:
+        yield f
+    finally:
+        fcntl.lockf(f, fcntl.LOCK_UN)
+        f.close()
 
 
 @contextmanager
@@ -55,15 +72,20 @@ def open_with_dirlock(*args, **kwargs):
     while no_attempt < MAX_ATTEMPTS:
         try:
             os.mkdir(lockfile)  # Atomic operation
-            f = open(*args, **kwargs)
-            yield f
-            f.close()
-            os.rmdir(lockfile)
             break
         except OSError:
-            logging.info("Can't immediately write-lock the file ({0}), retrying in {1} sec. ...".format(filename, TIME_BETWEEN_ATTEMPTS))
+            logging.info("Can't immediately write-lock the file ({0}), retrying in {1} sec.".format(filename, TIME_BETWEEN_ATTEMPTS))
             time.sleep(TIME_BETWEEN_ATTEMPTS)
             no_attempt += 1
+
+    if no_attempt == MAX_ATTEMPTS:
+        raise IOError("Failed to lock {0} {1} times.".format(filename, MAX_ATTEMPTS))
+
+    try:
+        with open(*args, **kwargs) as f:
+            yield f
+    finally:
+        os.rmdir(lockfile)
 
 
 def _fs_support_globalflock(fs):
